Move checkout components into folders
