+++ apps/store/src/app/[locale]/checkout/components/CheckoutForm/CheckoutForm.tsx
@@ -4,7 +4,7 @@ import { useApolloClient } from '@apollo/client'
 import { useTranslation } from 'next-i18next'
 import { useState } from 'react'
 import { Text, Button, BankIdIcon, yStack } from 'ui'
-import { useHandleSubmitCheckout } from '@/app/[locale]/checkout/useHandleSubmitCheckout'
+import { useHandleSubmitCheckout } from '@/app/[locale]/checkout/hooks/useHandleSubmitCheckout'
 import { useAsyncRouterPush } from '@/appComponents/useAsyncRouterPush'
 import * as FullscreenDialog from '@/components/FullscreenDialog/FullscreenDialog'
 import { PersonalNumberField } from '@/components/PersonalNumberField/PersonalNumberField'
@@ -22,9 +22,9 @@ import { useShopSession } from '@/services/shopSession/ShopSessionContext'
 import { useTracking } from '@/services/Tracking/useTracking'
 import { useRoutingLocale } from '@/utils/l10n/useRoutingLocale'
 import { PageLink } from '@/utils/PageLink'
+import { CheckoutStep, FormElement } from '../../CheckoutPage.constants'
+import { getCheckoutStepLink } from '../../CheckoutPage.helpers'
 import { errorPrompt } from './CheckoutForm.css'
-import { CheckoutStep, FormElement } from './CheckoutPage.constants'
-import { getCheckoutStepLink } from './CheckoutPage.helpers'
 
 type CheckoutFormProps = {
   shopSessionId: string
