+++ apps/store/src/app/[locale]/checkout/components/EmptyCart/EmptyCart.css.ts
